+++ package/ui/GroupRect.qml
@@ -18,21 +18,52 @@
 */
 
 import QtQuick 2.0
-import QtGraphicalEffects 1.0
+import QtGraphicalEffects 1.15
 
 import org.kde.plasma.plasmoid 2.0
 import org.kde.plasma.core 2.0 as PlasmaCore
 
-Item{
-    width: root.groupItemLength + 2 //layerWidth
+Item {
+    id: groupRoot
+    width: root.groupItemLength + 4 //layerWidth
     height: parent.height
-    clip: true
 
-    BackLayer{
-        anchors.right: parent.right
-        anchors.top: parent.top
+    Item {
+        anchors.fill: parent
 
-        width: 4*parent.width
-        height: parent.height
+        BackLayer {
+            id: secondRect
+            anchors.right: parent.right
+            anchors.top: parent.top
+            width: 4 * parent.width
+            height: parent.height
+        }
+
+        Rectangle {
+            id: mask
+            anchors.left: parent.left
+            anchors.top: parent.top
+            width: parent.width
+            height: parent.height
+            color: "transparent"
+            visible: false
+
+            Rectangle {
+                anchors.fill: parent
+                color: "black"
+
+                anchors.topMargin: 0.1 * parent.height
+                anchors.bottomMargin: anchors.topMargin
+                anchors.rightMargin: 6.8
+
+                radius: 6
+            }
+        }
+
+        layer.enabled: true
+        layer.effect: OpacityMask {
+            maskSource: mask
+            invert: true
+        }
     }
 }